# backend/openapi/openapi-v0.1.0.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: "0.1.0"
  description: |
    Контракты Backend/Frontend для таск-трекера.
    Документ ориентирован на использование в Swagger UI / Redoc и генерацию SDK.

    Базовый префикс всех эндпоинтов: /api/v1
servers:
  - url: http://localhost:8000/api/v1
    description: Основной сервер API
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/issues:
    get:
      operationId: listIssues
      tags: [Issues]
      summary: Список задач/эпиков с фильтрами
      description: |
        Основной эндпоинт для доски/таблицы задач.
        Поддерживает фильтры по статусу, исполнителю, автору, родителю, датам и поиску по ключу/заголовку.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: type
          in: query
          description: Тип сущности (epic или task). Если не передан — возвращаем оба.
          schema:
            $ref: "#/components/schemas/IssueType"
        - name: status_id
          in: query
          description: Фильтр по статусу.
          schema:
            $ref: "#/components/schemas/UUID"
        - name: assignee_id
          in: query
          description: Фильтр по исполнителю.
          schema:
            $ref: "#/components/schemas/UUID"
        - name: author_id
          in: query
          description: Фильтр по автору.
          schema:
            $ref: "#/components/schemas/UUID"
        - name: parent_id
          in: query
          description: Фильтр по родительской задаче (иерархия).
          schema:
            $ref: "#/components/schemas/UUID"
        - name: created_from
          in: query
          description: Нижняя граница даты создания (включительно).
          schema:
            $ref: "#/components/schemas/DateTime"
        - name: created_to
          in: query
          description: "Верхняя граница даты создания (см. реализацию: включительно/исключительно)."
          schema:
            $ref: "#/components/schemas/DateTime"
        - name: taken_in_work_from
          in: query
          description: Фильтр по дате взятия в работу.
          schema:
            $ref: "#/components/schemas/DateTime"
        - name: taken_in_work_to
          in: query
          description: Фильтр по дате взятия в работу.
          schema:
            $ref: "#/components/schemas/DateTime"
        - name: resolved_from
          in: query
          description: Фильтр по дате завершения.
          schema:
            $ref: "#/components/schemas/DateTime"
        - name: resolved_to
          in: query
          description: Фильтр по дате завершения.
          schema:
            $ref: "#/components/schemas/DateTime"
        - $ref: "#/components/parameters/Q"
        - name: sort
          in: query
          description: Поле сортировки списка.
          schema:
            type: string
            enum: [created_at, taken_in_work_at, resolved_at, due_date]
        - name: order
          in: query
          description: Направление сортировки.
          schema:
            type: string
            enum: [asc, desc]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Список задач
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedIssues"
    post:
      operationId: createIssue
      tags: [Issues]
      summary: Создать задачу/эпик
      description: |
        Создаёт сущность (epic/task). Поле key генерируется бэкендом (например PAY-123).
        Для задач onboarding и сигналов используйте единый источник правил:
        `backend/docs/onboarding_signals.md`.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
            examples:
              newEmployeeOnboardingMvpEpic:
                summary: Epic для MVP режима «Новый сотрудник»
                description: |
                  Epic фиксирует MVP-скоуп авто-onboarding:
                  - рекомендации по Wiki;
                  - рекомендации по Issues;
                  - рекомендации по ключевым людям (Users);
                  с объяснением «почему это важно» для каждого блока.
                value:
                  type: epic
                  title: Режим «Новый сотрудник» авто-onboarding (MVP)
                  description: Система автоматически рекомендует что читать (Wiki), какие задачи посмотреть (Issues), кто ключевые люди (Users). Только на основе данных проекта (активность, связи, роли), без вручную написанных гайдов.
                  status_id: "00000000-0000-0000-0000-000000000001"
              designOnboardingSignalsTask:
                summary: Task на проектирование сигналов onboarding (под эпиком)
                description: |
                  Создание task в эпике onboarding c parent_id=<EPIC_ID>.
                  Формулы/эвристики и формат ответа API описаны в:
                  `backend/docs/onboarding_signals.md`.
                value:
                  type: task
                  title: Спроектировать сигналы из реальных данных (что считаем важным)
                  description: Зафиксировать таблицы/факты, формулы важности и explainability (reason + score). См. backend/docs/onboarding_signals.md
                  status_id: "00000000-0000-0000-0000-000000000001"
                  parent_id: "00000000-0000-0000-0000-000000000099"
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema:
                type: object
                required: [issue]
                properties:
                  issue:
                    $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{projectId}/issues/by-key/{issueKey}:
    get:
      operationId: getIssueByKey
      tags: [Issues]
      summary: Получить задачу по ключу (например PAY-123)
      description: Удобно для быстрого поиска в UI по ключу задачи.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IssueKey"
      responses:
        "200":
          description: Задача
          content:
            application/json:
              schema:
                type: object
                required: [issue]
                properties:
                  issue:
                    $ref: "#/components/schemas/Issue"
        "404":
          $ref: "#/components/responses/NotFound"
  /issues/{issueId}:
    get:
      operationId: getIssue
      tags: [Issues]
      summary: Детальная карточка задачи
      description: |
        Возвращает задачу + связанные данные:
        - watchers (наблюдатели)
        - comments (комментарии)
        - attachments (вложения)
        - worklog_summary (план/факт по часам для карточки)
      parameters:
        - $ref: "#/components/parameters/IssueId"
      responses:
        "200":
          description: Детали
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetailsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateIssue
      tags: [Issues]
      summary: Обновить задачу (частично)
      description: |
        PATCH позволяет передавать только изменяемые поля.
        Бизнес-правила могут автоматически выставлять resolved_at при переводе в done.
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueRequest"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                type: object
                required: [issue]
                properties:
                  issue:
                    $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token в заголовке Authorization: Bearer <token>"
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице (пагинация).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка (пагинация).
      schema:
        type: integer
        minimum: 0
        default: 0
    Q:
      name: q
      in: query
      description: Поисковая строка (зависит от эндпоинта).
      schema:
        type: string
        maxLength: 200
...
    ProjectId:
      name: projectId
      in: path
      required: true
      description: UUID проекта.
      schema:
        $ref: "#/components/schemas/UUID"
    IssueId:
      name: issueId
      in: path
      required: true
      description: UUID задачи/эпика.
      schema:
        $ref: "#/components/schemas/UUID"
...
    IssueKey:
      name: issueKey
      in: path
      required: true
      description: Человекочитаемый ключ задачи (например PAY-123).
      schema:
        type: string
        pattern: "^[A-Z]{2,10}-\\d+$"
        example: PAY-123
...
  responses:
    BadRequest:
      description: Ошибка запроса / ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    UUID:
      type: string
      format: uuid
      description: UUID v4.
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    Date:
      type: string
      format: date
      description: Дата (YYYY-MM-DD).
      example: "2026-03-02"
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 datetime (с таймзоной).
      example: "2026-03-02T10:15:30Z"
    Numeric:
      type: number
      format: double
      description: Число (допускаются дробные значения).
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Поле запроса.
        reason:
          type: string
          description: Причина.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Код ошибки.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Текст ошибки.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
          description: ID запроса.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    UserLite:
      type: object
      required: [id, full_name]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        full_name:
          type: string
          description: Имя.
...
    IssueType:
      type: string
      enum: [epic, task]
      description: Тип задачи.
    Issue:
      type: object
      required: [id, project_id, key, type, title, status_id, author_id, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        project_id:
          $ref: "#/components/schemas/UUID"
        key:
          type: string
          description: Ключ задачи.
          example: PAY-123
        type:
          $ref: "#/components/schemas/IssueType"
        title:
          type: string
          description: Заголовок.
        description:
          type: string
          nullable: true
          description: Описание.
        status_id:
          $ref: "#/components/schemas/UUID"
          description: Статус.
        criticality_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
          description: Критичность.
        author_id:
          $ref: "#/components/schemas/UUID"
          description: Автор.
        assignee_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
          description: Исполнитель.
        parent_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
          description: Родитель (иерархия).
        start_date:
          $ref: "#/components/schemas/Date"
          nullable: true
          description: План начала.
        due_date:
          $ref: "#/components/schemas/Date"
          nullable: true
          description: Дедлайн.
        taken_in_work_at:
          $ref: "#/components/schemas/DateTime"
          nullable: true
          description: Взято в работу.
        resolved_at:
          $ref: "#/components/schemas/DateTime"
          nullable: true
          description: Завершено.
        planned_hours:
          $ref: "#/components/schemas/Numeric"
          nullable: true
          description: План часов.
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
    CreateIssueRequest:
      type: object
      required: [type, title, status_id]
      properties:
        type:
          $ref: "#/components/schemas/IssueType"
        title:
          type: string
        description:
          type: string
          nullable: true
        status_id:
          $ref: "#/components/schemas/UUID"
        criticality_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        assignee_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        parent_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        start_date:
          $ref: "#/components/schemas/Date"
          nullable: true
        due_date:
          $ref: "#/components/schemas/Date"
          nullable: true
        planned_hours:
          $ref: "#/components/schemas/Numeric"
          nullable: true
    UpdateIssueRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status_id:
          $ref: "#/components/schemas/UUID"
        criticality_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        assignee_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        parent_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        start_date:
          $ref: "#/components/schemas/Date"
          nullable: true
        due_date:
          $ref: "#/components/schemas/Date"
          nullable: true
        planned_hours:
          $ref: "#/components/schemas/Numeric"
          nullable: true
...
    IssueComment:
      type: object
      required: [id, issue_id, author_id, body, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        issue_id:
          $ref: "#/components/schemas/UUID"
        author_id:
          $ref: "#/components/schemas/UUID"
        body:
          type: string
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
...
    IssueAttachment:
      type: object
      required: [id, issue_id, uploaded_by, file_name, mime_type, size_bytes, storage_key, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        issue_id:
          $ref: "#/components/schemas/UUID"
        uploaded_by:
          $ref: "#/components/schemas/UUID"
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        storage_key:
          type: string
        created_at:
          $ref: "#/components/schemas/DateTime"
...
    WorklogSummary:
      type: object
      required: [planned_hours, logged_hours]
      properties:
        planned_hours:
          $ref: "#/components/schemas/Numeric"
          nullable: true
        logged_hours:
          $ref: "#/components/schemas/Numeric"
    IssueDetailsResponse:
      type: object
      required: [issue, watchers, comments, attachments, worklog_summary]
      properties:
        issue:
          $ref: "#/components/schemas/Issue"
        watchers:
          type: array
          items:
            $ref: "#/components/schemas/UserLite"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/IssueComment"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/IssueAttachment"
        worklog_summary:
          $ref: "#/components/schemas/WorklogSummary"
...
    PagedIssues:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total:
          type: integer
...